Remove AddLayer card
Since it was redundant with the hamburger menu. That menu's
icon has been replaced with the library_add icon used in the
AddLayer card.

Could not figure out a way to replace the icon with CSS, so
instead of resorting to a JavaScript solution I'm just
hiding the original hamburger icon and adding the new icon
next to it.