Show popup on each tile with GeoTIFF link
This link will allow the user to download a GeoTIFF of the
selected tile.

I tried to color the link with the color of the layer, and
while this was straightforward to do the first time, it was
hard to keep track of changing layer colors, so I've decided
to not do it at this time.

Also, the popups of the top-most layer are always active,
even if the layer is hidden, so to reach the underlying
layers the foremost layers have to be removed, which is less
than ideal.

But a good starting point.