Add checkboxes to sidebar
Previously any added layers would be removed from the
sidebar, and re-added if that layer was removed from
the list of active layers. However, this is confusing.
Thus, we make it so that all layers are always in the
sidebar, and add a checkbox that is checked if the
layer is currently on the map.

Also add a number of event handlers that clear the
layer properly whether it is removed from the card
or the sidebar.

This still needs styling improvements.

Closes #8